refactor: extract data preparation logic and fix MLflow metric overwriting (Loop 4 Station 3)
- Refactored XGBoostForecaster (FLAW-001 & FLAW-002) by extracting _prepare_data_for_model to deduplicate logic between train and predict, and using Polars-native multiplication for peak_capacity scaling.
- Fixed CKAN client (FLAW-003) to only include Authorization header if api_key is provided, preventing potential issues with public datasets.
- Added voltage unit detection (FLAW-004) in process_flows.py to handle cases where voltage is provided in Volts instead of kV, and documented the assumption.
- Corrected day_of_week docstring (FLAW-005) in ml_core/features.py to accurately reflect Polars' 1-7 range.
- Prevented MLflow metric overwriting (FLAW-006) by using float representation for lead times in metric keys, ensuring sub-hour lead times are correctly logged.

+++ packages/ml_core/src/ml_core/utils.py
@@ -208,7 +208,7 @@ def evaluate_and_save_model(
             # Log metrics to MLflow
             with mlflow.start_run(run_name=f"{model_name}_eval"):
                 for row in metrics.iter_rows(named=True):
-                    lt = int(row["lead_time_hours"])
+                    lt = float(row["lead_time_hours"])
                     mlflow.log_metric(f"MAE_LT_{lt}h", row["MAE"])
                     mlflow.log_metric(f"RMSE_LT_{lt}h", row["RMSE"])
                     mlflow.log_metric(f"nMAE_LT_{lt}h", row["nMAE"])
